The 44px floor moves out of cardCta.ts, because the join control is on the critical path (#2826)
`factionFaceSplit.test.ts` (#2079) caught the import the previous commit wrote:
`JoinControl` is reachable from `main.tsx` (through `InvitationLetterPopup`) and
from `pages/Settings.tsx` with no chunk boundary in between, and
`taskCard/cardCta.ts` sets eight faction `font-family` values. A static import of
that module from the entry graph strands the 62-rule faction font sheet — the
type paints in its generic fallback and nothing throws.

Geometry has no face, so the geometry splits off. `components/ctaTapFloor.ts`
holds `CARD_CTA` and nothing else; `cardCta.ts` re-exports it, so the eight
paints and all their consumers are untouched and there is still one name for one
box.

Refs #2826
